update: spotlight, videoOn, and hand raise vars to only update those vars
